async_api/pyproject.toml
[tool.poetry]
name = "async_api_sprint_1"
version = "0.1.0"
description = ""
authors = ["Ilsaf <ilsafganiev2222@gmail.com>", "Andrey <>"]

[tool.poetry.dependencies]
python = "^3.10"
aioredis = "^2.0.1"
elasticsearch = {version = "7.9.1", extras = ["async"]}
fastapi = "^0.79.0"
orjson = "^3.7.12"
pydantic = {extras = ["dotenv"], version = "^1.9.2"}
uvicorn = "^0.18.2"
elasticsearch-dsl = "^7.4.0"
gunicorn = "^20.1.0"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
flake8-bugbear = "^22.6.22"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.3.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-quotes = "^3.3.1"
flake8-simplify = "^0.19.2"
flake8-variables-names = "^0.0.5"
pre-commit = "^2.19.0"
dlint = "^0.13.0"
black = "^22.6.0"
isort = "^5.10.1"
yesqa = "^1.4.0"
flake8 = "^5.0.4"
pytest = "^7.1.2"
pytest-asyncio = "^0.19.0"
aiohttp = "^3.8.1"

[tool.black]
line-length = 120
target-version = ["py38"]
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
      \.venv
    | \.git
    | \.env
  )/
)
'''

[tool.isort]
multi_line_output = 3
line_length = 120
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
